dev_plan_agent/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-ai>=0.0.32
langgraph>=0.0.15
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0
requests>=2.28.2
langchain-core>=0.1.0
langsmith>=0.0.75
fastmcp>=0.4.1
